look for values of test.ping returns rather only the existence of the key
There seems to be a change on how test.ping returns 'down' minions.

Previously (salt 2016.x) it seemed to be fine to query the presence of
the key in the return. With at least salt-2018.x this is not true anymore

You get a return of {'node1': True, 'node2': False} (the entry remains in
the return of test.ping)

This entails that we always return True in 'minions.ready' although a node is
down.